Only owner can delete post issue fixed

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -141,7 +141,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId() + "/" + updatedImage.getTitle();
     }
 
 
@@ -149,9 +149,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirect) {
+        Image image = imageService.getImage(imageId);
+        User imgOwner = image.getUser();
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+
+        if(!imgOwner.getUsername().equalsIgnoreCase(loggedInUser.getUsername())){
+            String error = "Only the owner of the image can delete the image";
+            String imageTitle = image.getTitle();
+            redirect.addAttribute("deleteError", error).addFlashAttribute("deleteError", error);
+            return "redirect:images/"+imageId+"/"+imageTitle;
+        }
+        else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
